# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: E-Commerce Search API
  description: |
    Production-grade e-commerce search API powered by Elasticsearch 8.x.
    Supports full-text search, autocomplete, faceted filtering, multi-store, and analytics.
  version: 1.0.0
  contact:
    name: Search Team
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.example.com
    description: Production
...
paths:
...
  /api/search/suggest:
    get:
      tags: [Suggest]
      summary: Autocomplete suggestions
      description: |
        Returns autocomplete suggestions for queries, brands, categories, and products.
        Target response time: <100ms.
      operationId: suggest
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string, minLength: 1, maxLength: 200, example: "canon" }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 20, default: 10 }
        - name: store_id
          in: query
          schema: { type: integer }
        - name: types[]
          in: query
          schema:
            type: array
            items:
              type: string
              enum: [queries, brands, categories, products]
      responses:
        '200':
          description: Autocomplete suggestions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuggestResponse'
              example:
                queries: ["canon", "canon camera", "canon lens", "canon eos r6"]
                brands:
                  - { value: "Canon", label: "Canon", count: 342 }
                categories:
                  - { value: "Cameras", label: "Cameras", count: 156 }
                  - { value: "Mirrorless Cameras", label: "Mirrorless Cameras", count: 48 }
                products:
                  - id: 123
                    name: "Canon EOS R6 Mark II"
                    slug: "canon-eos-r6-mark-ii"
                    brand: "Canon"
                    price: 1999.99
                    rating: 4.8
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/search/click:
    post:
      tags: [Analytics]
      summary: Track product click
      description: Records when a user clicks on a product in search results for CTR analytics.
      operationId: trackClick
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query, product_id, position]
              properties:
                query: { type: string }
                product_id: { type: integer }
                position: { type: integer, minimum: 0 }
                store_id: { type: integer }
      responses:
        '200':
          description: Click tracked
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: "ok" }
...
components:
  securitySchemes:
    sanctum:
      type: http
      scheme: bearer
  schemas:
...
    SuggestResponse:
      type: object
      properties:
        queries:
          type: array
          items: { type: string }
        brands:
          type: array
          items:
            type: object
            properties:
              value: { type: string }
              label: { type: string }
              count: { type: integer }
        categories:
          type: array
          items:
            type: object
            properties:
              value: { type: string }
              label: { type: string }
              count: { type: integer }
        products:
          type: array
          items:
            type: object
            properties:
              id: { type: integer }
              name: { type: string }
              slug: { type: string }
              brand: { type: string }
              price: { type: number }
              rating: { type: number }
...
  responses:
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            type: object
            properties:
              message: { type: string }
              errors:
                type: object
                additionalProperties:
                  type: array
                  items: { type: string }